Complete all work

+++ HW4/Strassen_openmp.c
@@ -4,24 +4,24 @@
 #include <omp.h>
 
 #define INPUTFILE "input/test3"
-void malloc_matrix(int m, int n, unsigned ***matptr);
-void add(int m, int n, unsigned **mat1, unsigned **mat2, unsigned **mat3);
-void sub(int m, int n, unsigned **mat1, unsigned **mat2, unsigned **mat3);
+void malloc_matrix(int m, int n, double ***matptr);
+void add(int m, int n, double **mat1, double **mat2, double **mat3);
+void sub(int m, int n, double **mat1, double **mat2, double **mat3);
 
 // split one matrix to four same-shape sub-matrices
-void matrix_split(int m, int n, unsigned **mat,
-					unsigned **mat1, unsigned **mat2,
-					unsigned **mat3, unsigned **mat4);
+void matrix_split(int m, int n, double **mat,
+					double **mat1, double **mat2,
+					double **mat3, double **mat4);
 
 // merge four same-shape sub-matrices to one matrix
-void matrix_merge(int m, int n, unsigned **mat,
-					unsigned **mat1, unsigned **mat2,
-					unsigned **mat3, unsigned **mat4);
+void matrix_merge(int m, int n, double **mat,
+					double **mat1, double **mat2,
+					double **mat3, double **mat4);
 
-void multiply(int m1, int n1, unsigned **mat1, int m2, int n2, unsigned **mat2, unsigned **mat3);
-void multiply_NotP(int m1, int n1, unsigned **mat1, int m2, int n2, unsigned **mat2,unsigned **mat3);
+void multiply(int m1, int n1, double **mat1, int m2, int n2, double **mat2, double **mat3);
+void multiply_NotP(int m1, int n1, double **mat1, int m2, int n2, double **mat2,double **mat3);
 int thread_count=4;
-unsigned **A, **B, **C;
+double **A, **B, **C;
 
 int main(int argc,char*argv []) {
 	unsigned ma = 0;	
@@ -31,11 +31,11 @@ int main(int argc,char*argv []) {
 	int i = 0; // used in loop
 	int j = 0; // used in loop
 
-	unsigned **A11, **A12, **A21, **A22;
-	unsigned **B11, **B12, **B21, **B22;
-	unsigned **C11, **C12, **C21, **C22;
-	unsigned **C111, **C112, **C221, **C222;
-	unsigned **P1, **P2, **P, **Q1, **Q, **R1, **R, **S1, **S, **T1, **T, **U1, **U2, **U, **V1, **V2, **V;
+	double **A11, **A12, **A21, **A22;
+	double **B11, **B12, **B21, **B22;
+	double **C11, **C12, **C21, **C22;
+	double **C111, **C112, **C221, **C222;
+	double **P1, **P2, **P, **Q1, **Q, **R1, **R, **S1, **S, **T1, **T, **U1, **U2, **U, **V1, **V2, **V;
 	FILE *infile, *outfile;
 	
 	infile = fopen(INPUTFILE, "r");
@@ -50,7 +50,7 @@ int main(int argc,char*argv []) {
 	
 	for(i = 0; i < ma; i++)	// read A's content
 		for(j = 0; j < na; j++)
-			fscanf(infile, "%u", &A[i][j]);
+			fscanf(infile, "%lf", &A[i][j]);
 	
 	fscanf(infile, "%u %u", &mb, &nb);	// read B's dimensions
 	
@@ -64,7 +64,7 @@ int main(int argc,char*argv []) {
 	
 	for(i = 0; i < mb; i++)	// read B's content
 		for(j = 0; j < nb; j++)
-			fscanf(infile, "%u", &B[i][j]);
+			fscanf(infile, "%lf", &B[i][j]);
 	fclose(infile);
 	
 	
@@ -205,7 +205,7 @@ else if(argc>2&&atoi(argv[2])==3)
 	fprintf(outfile, "%d %d \n", ma, nb);
 	for(i = 0; i < ma; i++) {
 		for(j = 0; j < nb; j++)
-			fprintf(outfile, "%8.0u ", C[i][j]);
+			fprintf(outfile, "%8.0lf ", C[i][j]);
 		fprintf(outfile, "\n");	
 	}
 	fclose(outfile);
@@ -213,26 +213,26 @@ else if(argc>2&&atoi(argv[2])==3)
 	return 0;
 }
 
-void malloc_matrix(int m, int n, unsigned ***matptr) {
+void malloc_matrix(int m, int n, double ***matptr) {
 	/*
 	// the memory of the 2d array is not consecutive (each row)
 	int i;	
-	*matptr = malloc(m * sizeof(unsigned *));
+	*matptr = malloc(m * sizeof(double *));
 	for(i = 0; i < m; i++) 
-		(*matptr)[i] = malloc(n * sizeof(unsigned));
+		(*matptr)[i] = malloc(n * sizeof(double));
 	*/
 	
 	// the memory of the 2d array is consecutive (each row)
 	int i;
-	unsigned *tmp;	
-	tmp = (unsigned*)malloc(m * n * sizeof(unsigned *));
-	*matptr = (unsigned**)malloc(m * sizeof(unsigned *));
+	double *tmp;	
+	tmp = (double*)malloc(m * n * sizeof(double *));
+	*matptr = (double**)malloc(m * sizeof(double *));
 	for(i = 0; i < m; i++)
 		(*matptr)[i] = &(tmp[n*i]);
 	
 }
 
-void add(int m, int n, unsigned **mat1, unsigned **mat2, unsigned **mat3) {	
+void add(int m, int n, double **mat1, double **mat2, double **mat3) {	
 	int i, j;
 	#pragma omp parallel for num_threads(thread_count) private(i,j) shared(mat3,mat2,mat1)
 	for(i = 0; i < m; i++)
@@ -240,7 +240,7 @@ void add(int m, int n, unsigned **mat1, unsigned **mat2, unsigned **mat3) {
 			mat3[i][j] = mat1[i][j] + mat2[i][j];	
 }
 
-void sub(int m, int n, unsigned **mat1, unsigned **mat2, unsigned **mat3) {	
+void sub(int m, int n, double **mat1, double **mat2, double **mat3) {	
 	int i, j;
 	#pragma omp parallel for num_threads(thread_count) private(i,j) shared(mat3,mat2,mat1)
 	for(i = 0; i < m; i++)
@@ -248,7 +248,7 @@ void sub(int m, int n, unsigned **mat1, unsigned **mat2, unsigned **mat3) {
 			mat3[i][j] = mat1[i][j] - mat2[i][j];	
 }
 
-void multiply(int m1, int n1, unsigned **mat1, int m2, int n2, unsigned **mat2, unsigned **mat3) {	
+void multiply(int m1, int n1, double **mat1, int m2, int n2, double **mat2, double **mat3) {	
 	int i, j, k;
 
 #pragma omp parallel for num_threads(thread_count) collapse(2)  private(i,j) shared(mat3)
@@ -264,7 +264,7 @@ void multiply(int m1, int n1, unsigned **mat1, int m2, int n2, unsigned **mat2, 
 
 
 }	
-void multiply_NotP(int m1, int n1, unsigned **mat1, int m2, int n2, unsigned **mat2, unsigned **mat3) {	
+void multiply_NotP(int m1, int n1, double **mat1, int m2, int n2, double **mat2, double **mat3) {	
 	int i, j, k;
 
 	for(i = 0; i < m1; i++)
@@ -277,9 +277,9 @@ void multiply_NotP(int m1, int n1, unsigned **mat1, int m2, int n2, unsigned **m
 				mat3[i][k] += mat1[i][j] * mat2[j][k];
 }				
 
-void matrix_split(int m, int n, unsigned **mat,
-					unsigned **mat11, unsigned **mat12,
-					unsigned **mat21, unsigned **mat22) {
+void matrix_split(int m, int n, double **mat,
+					double **mat11, double **mat12,
+					double **mat21, double **mat22) {
 	int i, j;
 	if(m % 2 != 0 || n % 2 != 0) {
 		printf("m = %d, n = %d \n", m, n);
@@ -308,9 +308,9 @@ void matrix_split(int m, int n, unsigned **mat,
 	
 }
 
-void matrix_merge(int m, int n, unsigned **mat,
-					unsigned **mat11, unsigned **mat12,
-					unsigned **mat21, unsigned **mat22) {
+void matrix_merge(int m, int n, double **mat,
+					double **mat11, double **mat12,
+					double **mat21, double **mat22) {
 	int i, j;
 	if(m % 2 != 0 || n % 2 != 0) {
 		printf("error, cannot merge by 4 sub-matrices \n");
